backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
Pillow==10.1.0
requests==2.31.0
setuptools==65.5.0
kafka-python==2.0.2

